requirements.txt
google-cloud-vision
google-generativeai
pandas
openpyxl
Pillow
